+++ bugsnag-android-core/src/main/java/com/bugsnag/android/UserRepository.kt
@@ -13,13 +13,7 @@ internal class UserRepository(private val prefs: SharedPreferences, private val 
     }
 
     fun load(): User {
-        var installId = prefs.getString(INSTALL_ID_KEY, null)
-
-        if (installId == null) {
-            installId = UUID.randomUUID().toString()
-            prefs.edit().putString(INSTALL_ID_KEY, installId).apply()
-        }
-
+        val installId = getDeviceId()
         return when {
             persist -> // Check to see if a user was stored in the SharedPreferences
                 User(
@@ -45,5 +39,14 @@ internal class UserRepository(private val prefs: SharedPreferences, private val 
                 .remove(USER_EMAIL_KEY)
         }
         editor.apply()
+    }
+
+    fun getDeviceId(): String {
+        var installId = prefs.getString(INSTALL_ID_KEY, null)
+        if (installId == null) {
+            installId = UUID.randomUUID().toString()
+            prefs.edit().putString(INSTALL_ID_KEY, installId).apply()
+        }
+        return installId
     }
 }